backend/requirements.txt
cryptography
fastapi
pydantic
pymongo
python-dotenv
uvicorn
requests
agno>=0.1.0
